+++ lan/src/ne.c
@@ -252,7 +252,7 @@ int er$init_tables() {
   return SS$_NORMAL;
 }
 
-long er_iodb_vmsinit(int dev) {
+long er_iodb_vmsinit(long dev) {
 #if 0
   struct _ucb * ucb=&er$ucb;
   struct _ddb * ddb=&er$ddb;
@@ -306,7 +306,7 @@ long er_iodbunit_vmsinit(struct _ddb * ddb,int unitno,void * dsc) {
 
 extern char * mydevice;
 
-int er_vmsinit(int dev) {
+int er_vmsinit(long dev) {
   //struct _ucb * u=makeucbetc(&ddb,&ddt,&dpt,&fdt,"hda","hddriver");
 
   unsigned short chan0, chan1, chan2;
